Add native diff view to SwiftUI client
- Adds Phase-2 diff-view flow to the macOS SwiftUI client: selecting a changed file now shows its diff with syntax colour, intra-line highlights, and +/- totals, mirroring the Tauri client's three-step get_diff -> parse_diff -> tokenize_diff pipeline
- Introduces DiffView, DiffLineText, DiffPalette, and DiffStore in the Swift app, plus a purpose-built DiffPayload record and new UniFFI-exported diff functions in the FFI crate
- Makes core::highlight::tokenize_diff public and switches IntraLineRange to u32 indices so it matches Token and stays UniFFI-representable
- Updates DESIGN.md, README.md, ROADMAP.md, and TECHNICAL.md to reflect the new diff-view flow and narrow near-term scope to macOS-app-only

+++ core/src/render.rs
@@ -125,8 +125,8 @@ fn render_slice(
         emit_span(&chars[start..end], base_class, out);
         return;
     };
-    let overlap_start = start.max(intra.start);
-    let overlap_end = end.min(intra.start + intra.length);
+    let overlap_start = start.max(intra.start as usize);
+    let overlap_end = end.min((intra.start + intra.length) as usize);
     if overlap_start >= overlap_end {
         emit_span(&chars[start..end], base_class, out);
         return;
